requirements.txt
Django>=4.0
djangorestframework
django-filter
django-import-export
django-simple-history
Pillow
openpyxl
flake8